data_cleaning/requirements.txt
# Function dependencies, for example:
# package>=version
nltk
google-cloud-storage